Update avc denials returned by test so:
only denials related to given test are counted (and not all from all tests)
information about denials are summarized by process so the Error in jenkis is unified and same errors are marked as one not distinguished by pids, randomized file names etc.
new provider, that is able to create such summary introduced, so it can be used also in other cases if necessary
Reason:

actual test is reporting more than 70 denails even there are only 9 of them.
Failures with same reason are reported in Jira as single tasks, and they can be logged under same.