filesystem_device: improve parameter coverage
Cover more test values for parameters according to our test plan.

1.) thread_pool_size = 0
2.) sandbox mode

Furthermore, the test constructed a long regular expression assuming a certain
order of the parameters to be present in the virtiofsd command line.
However, the order is not relevant, so instead of constructing one long
regular expression add one for each parameter instead.
This way the test won't fail if the order was different and it's easier
to add new expectations.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>