filesystem_device: sync shared folder to avoid caching problems
Sometimes md5sum gets stuck. This might happen because there's a problem
with the file. In each guest sync the mounted folder before creating the
md5sum to make sure the cache and storage are synchronized.

Also, move the filename definition out because it was only defined once
in the first loop giving the wrong impression that it would vary.
However, we always write only one file so it's reasonable to have a
single file name.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>