Remove 2 checkpoints for passt
Delete 2 checkpoints:
1) Check the socket dir, when vm destroyed, the passt socket dir is not
   cleared for passt+vhostuser. In CNV, this is not an issue, tracked in
RHEL-80285. (We can restore the check after the bug is fixed)
2) When start vm with interface using passt, and the qemu support
   reconect parameter, there is "reconnect:5" or "reconnect=5" in qemu
cmd once qemu support reconnect feature. But since qemu 10.1, the
parameter change to "reconnect-ms:5000" or "reconnect-ms=5000". To make
is simple, delete this qemu cmd check, and just check the function that
if the passt process will reconnect after killed is enough.

Signed-off-by: Yalan Zhang <yalzhang@redhat.com>