passt: support multiple nexthops
In some environment, there are more than one gateway in nexthops in ip route
output, for example on ARM host.
This patch is to support this situation instead of only
selecting the first default route which is not accurate.

For example:
"nexthops":[{"gateway":"fe80::4a5a:d00:6631:2920","dev":"eno1","weight":1,"flags":[]},
            {"gateway":"fe80::4a5a:d00:6631:b20","dev":"eno1","weight":1,"flags":[]}
           ]

Signed-off-by: Dan Zheng <dzheng@redhat.com>