[virsh metadata] change the checking logic for missing metadata
from libvirt 11.3.0, instead of return status error, libvirt will return an empty string if there is no metadata to be returned
Added a new parameter 'metadata_empty' for the negative test cases. This parameter is used to verify that the metadata were not altered by the preceding command.

refer to
https://gitlab.com/libvirt/libvirt/-/commit/312088d9b62334c950310f369b2e0fe2302a74a3
--- tools: virsh: metadata: do not report error on missing metadata

Signed-off-by: zhentang-tz <zhetang@redhat.com>