Delete a negative test with invalid setting
There was a negative test to save with formats as "abc". The behavior changes to handel such kind of settings in qemu.conf. Before, the setting can be saved and the daemon can start, but virsh save failed; Now, the daemon can not start with such invalid setting. Delete this negative test as it can not cause any problem if daemon can not start with it.

Signed-off-by: Yalan Zhang <yalzhang@redhat.com>