virsh_dump: handle config error
In more recent versions of libvirt, the qemu.conf entry for the
compression algorithm is validated. In that case the daemon won't start.

Change the test case so that it bails out if it finds the entry was
validated. If it isn't, continue as before.

Don't use here the libvirt version as a switch because it could lead to
problems if the validation behavior was not the same as the upstream
version when it was introduced.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>