guest_agent: Domfsfreeze domfsthaw fix (#6549)
* domfsfreeze_domfsthaw: Add a fail pattern for frozen state check

When vm is in frozen state, running a agent command should fail
with clear error message. The error message varies depending on
the libvirt and qemu versions. The previos fail patterns misses
the error message on RHEL9.0

* domfsfreeze_domfsthaw: Thaw vm filesystems before cleanup

If the test fails between domfsfreeze and domfsthaw, cleanup will
fail because it tries to remove files in vm when vm is frozen.