controller_functional: fix test for s390x
'generic-ccw' was the standard device name in lszdev in the past
but now it's been changed to actually list the virtio device type.
However, this test might have to run on older versions still and
the test expectation is only that the device is listed for the given
device number. So, omit the device type restriction to check all
devices. with s390-virtio, there are only virtio devices listed with
lszdev (or passed through devices but they are not present here / per
default).

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>