migration_xml_update: set mac to avoid connectivity issue
The update_interface_mtu test case has been failing because the IP cache
was out of date. Although the VM can be logged into the test failed to
continue because the cache tried to access the IP through an old mac
address.

However, when the test modifies the interface and it doesn't require a
mac libvirt will assign some mac address which might be different from
the previous one. This leads to the framework's cache to become invalid.
Instead, now update the xml so that there's no mismatch between XML and
the mac and the connection can be established.

This seems to have worked when the test case was submitted recently but
I'd attribute this to a coincidence. Recent test runs on s390x and
x86_64 all fail due to this failure of connectivity from the test
framework with error message

"""
virttest.virt_vm.VMAddressVerificationError: Could not verify DHCP
lease: 52:54:00:8b:b2:59 --> ...
"""

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>