Replacing pidof with pgrep to make it more generic!
In the latest versions of QEMU, the QEMU process may run as qemu-system-<arch>, where <arch> represents any supported architecture. However, in some cases, it may still run as qemu-kvm. To ensure a more generic and reliable approach for identifying the QEMU process ID, it is recommended to replace the pidof command with pgrep, as pgrep provides broader compatibility and flexibility.
Signed-off-by: Anushree-Mathur <anushree.mathur@linux.ibm.com>