virtual_network/migrate_with_bridge: enable remote host iface config
We previously enabled configuration for host face selection for bridge
setup to account for environments where it's not the default interface
(the first one). However, the test case also requires this for the
remote host as it's a migration test case.

Add this option and the same setting logic as for 'host_iface'.
For this, the setting logic has to be brought further up in the
execution where the variable is defined.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>