virtual_disks/dasd: improve config state detection
Previously, the code relied on an error message to decide if a device
was online already and if so, not set it offline during teardown.

However, this message wasn't shown always; possibly because of https://issues.redhat.com/browse/RHEL-90213

Instead, use the lszdev tool to detect if the device is already online.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>